+++ docs/scripts/index-static.js
@@ -4,21 +4,13 @@ import pkg from "../package.json"
 import routes from "app/routes"
 import store from "app/store"
 
-module.exports = ({
-  urls,
-  pagesData,
-  dest,
-  baseUrl,
-}) => (
+module.exports = (options) => (
   statinamicStatic({
+    ...options,
+    routes,
+    store,
     metadata: {
       pkg,
     },
-    urls,
-    pagesData,
-    dest,
-    baseUrl,
-    routes,
-    store,
   })
 )